+++ changelog.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- Change ``AsyncPlayerPreLoginEvent`` to ``PlayerLoginEvent``
+
 # 1.0.0.2
 
 - Added the command `/ess convert EssentialsX`, allows converting the database from EssentialsX to zEssentials.